Split generated tooling into dedicated libraries

+++ Samples/Demos/BusinessDataVaultSql/cleanup.cmd
@@ -1,2 +1,2 @@
 cd /d C:\Users\jimmy\Desktop\meta-bi\Samples\Demos\BusinessDataVaultSql
-rmdir /s /q GeneratedSql
+if exist GeneratedSql rmdir /s /q GeneratedSql
